Allow using supertabular instead of longtable in listings
Longtable is not supported in two columns mode, this is a workaround
to avoid this problem.

+++ verif/tests/actions/load_combinations/comb_container/test_comb_container_03.py
@@ -25,6 +25,7 @@ texFileName= os.path.basename(texFileNameRef)
 texFileName= texFileName.replace('ref_','./')
 ## Write LaTeX file
 combContainer.exportToLatex(fileName= texFileName)
+# combContainer.exportToPDF(fileName= texFileName.replace('.tex', '.pdf'), superTabular= False)
 
 # ## Generate PDF.
 # pdfFileName= csvFileName.replace('.csv','.pdf')
@@ -39,5 +40,6 @@ if comparisonOK:
    print('test '+fname+': ok.')
 else:
     lmsg.error(fname+' ERROR.')
-os.remove(texFileName) # Your garbage, you clean it.
+if(comparisonOK):
+    os.remove(texFileName) # Your garbage, you clean it.
